Clarify TTM alias and share metric logic
